+++ internal/session/helpers_test.go
@@ -1,0 +1,7 @@
+package session_test
+
+import "os"
+
+func statFile(path string) (os.FileInfo, error) {
+	return os.Stat(path)
+}